Add custom workout

+++ frontend/src/components/WorkoutDetails.js
@@ -1,13 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+    if (!user) {
+      return 
+    }
     const response = await fetch('/api/workouts/' + workout._id, {
-      method: 'DELETE'
+      method: 'DELETE',
+      headers: {
+        'authorization': `Bearer ${user.token}`
+      }
     })
     const json = await response.json()
 
@@ -20,7 +28,7 @@ const WorkoutDetails = ({ workout }) => {
     <div className="workout-details">
       <h4>{workout.title}</h4>
       <p><strong>Load (kg): </strong>{workout.load}</p>
-      <p><strong>Reps (kg): </strong>{workout.reps}</p>
+      <p><strong>Reps: </strong>{workout.reps}</p>
       <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
       <span onClick={handleClick}><i className="bi bi-trash"></i></span>
     </div>
